Move the Dockerfile to the root and split the dev servers
The image build already needed the whole workspace as its context, so the
Dockerfile sits beside docker-compose.yml, which now picks it up as the context
default.

pnpm dev runs the API and Vite in parallel; dev:server no longer builds the UI,
since Vite serves it in dev, and dev:server:ui covers the one-origin flow.
FRONTEND_PORT moves Vite off :3000.

Frontend tests move to packages/frontend/tests/, mirroring src/. In frontend.ts
"shell" becomes index.html, with a comment on why client-side routes fall back
to it instead of 404ing.
