+++ package.json
@@ -12,9 +12,11 @@
     "format:fix": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\" --cache --config .prettierrc",
     "test": "pnpm --filter @truefoundry/utils test && pnpm --filter @truefoundry/server test && pnpm --filter frontend test",
     "test:store:local": "bash scripts/test-store-with-postgres.sh",
-    "dev:server": "pnpm --filter frontend build && pnpm --filter @truefoundry/server dev",
-    "openapi:write": "pnpm --filter @truefoundry/server openapi:write",
+    "dev": "pnpm --parallel --filter @truefoundry/server --filter frontend dev",
+    "dev:server": "pnpm --filter @truefoundry/server dev",
     "dev:frontend": "pnpm --filter frontend dev",
+    "dev:server:ui": "pnpm --filter frontend build && pnpm dev:server",
+    "openapi:write": "pnpm --filter @truefoundry/server openapi:write",
     "ui:shot": "node ui-shots/shot.mjs",
     "ui:clean": "node ui-shots/clean.mjs",
     "pack:dry": "pnpm --filter @truefoundry/utils pack:dry",
