+++ packages/frontend/package.json
@@ -11,7 +11,7 @@
     "build": "tsc --noEmit && vite build",
     "preview": "vite preview",
     "typecheck": "tsc --noEmit",
-    "test": "tsx --test \"src/**/*.test.ts\""
+    "test": "tsx --test \"tests/**/*.test.ts\""
   },
   "dependencies": {
     "@assistant-ui/core": "0.2.21",
