+++ packages/frontend/tsconfig.json
@@ -20,6 +20,6 @@
     "isolatedModules": true,
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.tsbuildinfo"
   },
-  "include": ["src"],
-  "exclude": ["src/**/*.test.ts"]
+  "include": ["src", "tests"],
+  "exclude": ["tests/**/*.test.ts"]
 }
